Escaping '.' on English singular units.
The '.' character is interpreted as a Meta Escape when passed to the regex which matches these units. One of the problems this causes is that the abreviation for  pounds 'lb' is becomes a positive match for liters 'l.' because the regex ends up as /(?=\\bl.\\b)/ which is interpreted as a 2-letter word beginning with the letter 'l' and ending with any letter.